+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Patching support and docs enhancements were added for the ConfigMap resource.
+
 ## [0.2.47] - 2025-04-07
 
 ### Fixed